修复部分已知bug，新增获取二维码功能

+++ DWeChatRobot/ForwardMessage.h
@@ -1,6 +1,6 @@
 #pragma once
 #include <windows.h>
-BOOL __stdcall ForwardMessage(wchar_t *wxid, int localId);
+BOOL __stdcall ForwardMessage(wchar_t *wxid, unsigned long long localId);
 
 #ifndef USE_SOCKET
 extern "C" __declspec(dllexport) BOOL ForwardMessageRemote(LPVOID lpParameter);
